refactor(providers): invert provider→application handler and native-spec deps
Provider implementations no longer import from orb.application. The
provider-handler contract and its shared template-method base
(ProviderHandler / BaseProviderHandler) now live in
orb.domain.base.provider_handler_base; the application-layer modules
re-export them for backward compatibility. AWS handlers inherit the base
from the domain layer directly.

The generic native-spec service is now resolved through a new domain
NativeSpecPort. The concrete application NativeSpecService is registered
under the port during DI bootstrap, so the AWS and Kubernetes provider
wiring resolves it via the port instead of importing the application
service, keeping the dependency direction pointing inward.
