+++ src/orb/application/interfaces/provider_handler.py
@@ -1,42 +1,11 @@
-"""Provider Handler interface for CQRS pattern consistency."""
+"""Provider Handler interface for CQRS pattern consistency.
 
-from abc import ABC, abstractmethod
-from typing import Generic, Optional, TypeVar
+The canonical definition now lives in the domain layer
+(:mod:`orb.domain.base.provider_handler_base`) so that provider
+implementations can depend on it without importing upward into the
+application layer.  It is re-exported here for backward compatibility.
+"""
 
-TRequest = TypeVar("TRequest")
-TResponse = TypeVar("TResponse")
+from orb.domain.base.provider_handler_base import ProviderHandler
 
-
-class ProviderHandler(Generic[TRequest, TResponse], ABC):
-    """
-    Base interface for provider handlers.
-
-    This interface provides the contract for all provider handlers in the system,
-    following the same architectural pattern as other handler interfaces:
-
-    - ProviderHandler (interface) → BaseProviderHandler (implementation)
-    - CommandHandler (interface) → BaseCommandHandler (implementation)
-    - QueryHandler (interface) → BaseQueryHandler (implementation)
-    - EventHandler (interface) → BaseEventHandler (implementation)
-    - InfrastructureHandler (interface) → BaseInfrastructureHandler (implementation)
-
-    This ensures architectural consistency across all handler types in the CQRS system.
-    """
-
-    @abstractmethod
-    async def handle(self, request: TRequest, context: Optional[object] = None) -> TResponse:
-        """
-        Handle a provider request.
-
-        Args:
-            request: Provider request to handle (strongly typed)
-            context: Optional provider context
-
-        Returns:
-            Provider response (strongly typed)
-
-        Raises:
-            ValidationError: If request is invalid
-            ProviderError: If provider operation fails
-            InfrastructureError: If infrastructure operation fails
-        """
+__all__ = ["ProviderHandler"]
